+++ CBV.h
@@ -196,8 +196,8 @@ public:
 
 //******************** Operations of weighting, finding and casing **********************
   int CountBit() const;
-  int LeftOne(int nNext = -1) const;
-  int LeftOne(BYTE& bt) const;
+  ptrdiff_t LeftOne(ptrdiff_t nNext = -1) const;
+  ptrdiff_t LeftOne(BYTE& bt) const;
   int RightOne(int nNext = -1) const;
   int RightOne(BYTE& bt) const;
